fix payment return data

+++ ManagementSchool/Service/TuitionFeeNotificationService/ITuitionFeeNotificationService.cs
@@ -5,7 +5,7 @@ namespace ManagementSchool.Service.TuitionFeeNotificationService;
 
 public interface ITuitionFeeNotificationService
 {
-    Task<bool> CreateTuitionFeeNotificationAsync(string semesterType, string academicYear, decimal amount, string content);
-    Task<bool> UpdateTuitionFeeNotificationAsync(SemesterType semesterType, string academicYear, decimal amount, string content);
+    Task<bool> CreateTuitionFeeNotificationAsync(string semesterType, string academicYear, double amount, string content);
+    Task<bool> UpdateTuitionFeeNotificationAsync(SemesterType semesterType, string academicYear, double amount, string content);
     Task<TuitionFeeNotificationDto> GetTuitionFeeNotificationAsync(SemesterType semesterType, string academicYear);
 }
